move processing of labels and annotations as tags
