+++ internal/controller/datadogagent/feature/enabledefault/rbac_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package override
+package enabledefault
 
 import (
 	"testing"
